requirements-dev.txt
# Include production requirements
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
factory-boy>=3.3.0
faker>=20.1.0
selenium>=4.16.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0
mypy>=1.7.0
django-stubs>=4.2.7

# Security
bandit>=1.7.5
safety>=2.3.5

# Development Tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
ipython>=8.18.0
ipdb>=0.13.13

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=2.0.0
